+++ packages/col-picker/src/main.vue
@@ -42,6 +42,7 @@
       :title="title || t('wd.colPicker.title')"
       :close-on-click-modal="closeOnClickModal"
       :close-on-popstate="closeOnPopstate"
+      :safe-area-inset-bottom="safeAreaInsetBottom"
       @close="handlePickerClose"
     >
       <div class="wd-col-picker__selected" ref="selectedVisible">
@@ -167,6 +168,10 @@ export default {
     },
     autoComplete: Boolean,
     closeOnPopstate: {
+      type: Boolean,
+      default: true
+    },
+    safeAreaInsetBottom: {
       type: Boolean,
       default: true
     }